- Move generateFilterTags to utils - Move tags logic to ApiTags component
